+++ tcp/mux.go
@@ -39,41 +39,18 @@ func init() {
 // make connections to other nodes, and receive connections from other
 // nodes.
 type Layer struct {
-	ln     net.Listener
-	header byte
-	addr   net.Addr
+	ln   net.Listener
+	addr net.Addr
 
-	remoteEncrypted bool
-	skipVerify      bool
-	nodeX509CACert  string
-	tlsConfig       *tls.Config
+	dialer *Dialer
+
+	nodeX509CACert string
+	tlsConfig      *tls.Config
 }
 
 // Dial creates a new network connection.
 func (l *Layer) Dial(addr string, timeout time.Duration) (net.Conn, error) {
-	dialer := &net.Dialer{Timeout: timeout}
-
-	var err error
-	var conn net.Conn
-	if l.remoteEncrypted {
-		conf := &tls.Config{
-			InsecureSkipVerify: l.skipVerify,
-		}
-		conn, err = tls.DialWithDialer(dialer, "tcp", addr, conf)
-	} else {
-		conn, err = dialer.Dial("tcp", addr)
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	// Write a marker byte to indicate message type.
-	_, err = conn.Write([]byte{l.header})
-	if err != nil {
-		conn.Close()
-		return nil, err
-	}
-	return conn, err
+	return l.dialer.Dial(addr, timeout)
 }
 
 // Accept waits for the next connection.
@@ -263,14 +240,12 @@ func (mux *Mux) Listen(header byte) *Layer {
 	mux.m[header] = ln
 
 	layer := &Layer{
-		ln:              ln,
-		header:          header,
-		addr:            mux.addr,
-		remoteEncrypted: mux.remoteEncrypted,
-		skipVerify:      mux.InsecureSkipVerify,
-		nodeX509CACert:  mux.x509CACert,
-		tlsConfig:       mux.tlsConfig,
+		ln:             ln,
+		addr:           mux.addr,
+		nodeX509CACert: mux.x509CACert,
+		tlsConfig:      mux.tlsConfig,
 	}
+	layer.dialer = NewDialer(header, mux.remoteEncrypted, mux.InsecureSkipVerify)
 
 	return layer
 }