+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Implementation changes and bug fixes
 - [PR #857](https://github.com/rqlite/rqlite/pull/857): Use Protobufs as core data model.
+- [PR #858](https://github.com/rqlite/rqlite/pull/858): Create dedicated client for talking to a cluster service.
 
 ## 6.1.0 (August 5th 2021)
 This release makes significant changes to SQLite database connection handling, resulting in proper support for high-performance concurrent reads of in-memory databases (an in-memory SQLite database is the default option for rqlite).